Some Changes request by Lucas solved
