--- Assets/Scripts/Core/Battle.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 60583e9bbe6966647b3cf7ecb718bfb3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
