--- Assets/Scripts/Core/Battle/CustomLookAtTarget.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8617eced157a5fc4cb6ccf3b380f4709
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
