+++ Assets/Scripts/Core/Characters/Marine.cs
@@ -1,4 +1,5 @@
 using DG.Tweening;
+using System.Collections;
 using UnityEngine;
 using UnityEngine.AI;
 
@@ -10,7 +11,6 @@ public class Marine : MonoBehaviour
     public Animator animator;
     public AnimatorOverrideController animatorOverride;
     public SelectableCharacter characterSelectionImage;
-    public CustomLookAtTarget customLookAtTarget;
     public WeaponBase actualWeapon;
     [SerializeField] private GameObject selectionArrow;
 
@@ -28,6 +28,10 @@ public class Marine : MonoBehaviour
     public Walking WalkingState { get; set; }
     public Shooting ShootingState { get; set; }
     public Reloading ReloadingState { get; set; }
+
+    [Header("LookAt")]
+    [SerializeField] private float speed = 1f;
+    public Coroutine LookCoroutine { get; private set; }
 
     private void Awake()
     {
@@ -87,5 +91,29 @@ public class Marine : MonoBehaviour
         //No hacer destroy, dejar el cuerpo ahi tirado en el piso
         //Quitar de la lista de myMairnes
         Destroy(gameObject);
+    }
+
+    public void StartRotating(Transform target)
+    {
+        if (LookCoroutine != null)
+        {
+            StopCoroutine(LookCoroutine);
+        }
+
+        LookCoroutine = StartCoroutine(LookAt(target));
+    }
+
+    private IEnumerator LookAt(Transform target)
+    {
+        Quaternion lookRotation = Quaternion.LookRotation(target.position - transform.position);
+
+        float time = 0;
+
+        while (time < 1)
+        {
+            transform.rotation = Quaternion.Slerp(transform.rotation, lookRotation, time);
+            time += Time.deltaTime * speed;
+            yield return null;
+        }
     }
 }
